Changed AutoVectorize to use return by value
It's a proof of concept that shows that the auto vectorizer can create
similar performance than the manual SSE and ISPC.

The issue with the return by ref is that the optimizer is bad at
making assumptions about memory. Also in general the optimizer is very good
at widening but doing so manually might irritate him more.

Also changed to -O1 for main.cpp. It is mostly required to get rid of tuple boilerplate.
And return by value by itself benefits from having optimizations for the Caller.
Checked the ASM: AutoVectorize doesn't get inlined and gets no possible benefit this way.